create team form is completely moved to own component, removed all props expect one for errors
