+++ src/components/CreateTeam/CreateTeam.jsx
@@ -1,9 +1,40 @@
 import { Button, Box, Typography, Paper, TextField, Grid, FormLabel} from '@mui/material';
+import {useSelector, useDispatch} from 'react-redux';
+import { useEffect, useState } from 'react';
 
 function CreateTeam(props) {
+    const dispatch = useDispatch();
+    const [teamName, setTeamName] = useState('');
+    const [league, setLeague] = useState('');
+    const [season, setSeason] = useState('');
+    const [yourPlayerNumber, setYourPlayerNumber] = useState('');
+    
+    const registerTeam = (event) => {
+        event.preventDefault();
+        // Send data to saga for POST to server
+        dispatch({
+          type: 'ADD_TEAM',
+          payload: {
+            teamName,
+            league,
+            year: season,
+            number: yourPlayerNumber
+          }
+        });
+        clearCreateTeam();
+      }; // end registerTeam
+
+    // Clear the local state for create team form
+    const clearCreateTeam = () => {
+    setTeamName('');
+    setLeague('');
+    setSeason('');
+    setYourPlayerNumber('');
+  }
+
     return (
         <Paper elevation={8} sx={{mb: 1, minWidth: '300px', width: '80%'}}>
-        <form onSubmit={props.registerTeam}>
+        <form onSubmit={registerTeam}>
           <Grid container alignItems='center' justify='center' direction='column'>
           <Typography variant='h4' gutterBottom>Add Team</Typography>
           {props.errors.registrationMessage && (
@@ -11,9 +42,6 @@ function CreateTeam(props) {
               {props.errors.registrationMessage}
             </Typography>
           )}
-          {props.teamPlayers.allTeams.length > 0 && props.teamPlayers.allTeams.map((team, index) => {
-            return <p key={index}>{team.name} ID #{team.id}</p>
-          })}
           <Grid item sx={{mb: 1}}>
             <FormLabel htmlFor="TeamName">
               <TextField
@@ -23,9 +51,9 @@ function CreateTeam(props) {
                 type="text"
                 name="TeamName"
                 filled='true'
-                value={props.teamName}
+                value={teamName}
                 required
-                onChange={(event) => props.setTeamName(event.target.value)}
+                onChange={(event) => setTeamName(event.target.value)}
               />
             </FormLabel>
           </Grid>
@@ -38,9 +66,9 @@ function CreateTeam(props) {
                 filled='true'
                 type="text"
                 name="league"
-                value={props.league}
+                value={league}
                 required
-                onChange={(event) => props.setLeague(event.target.value)}
+                onChange={(event) => setLeague(event.target.value)}
               />
             </FormLabel>
           </Grid>
@@ -53,9 +81,9 @@ function CreateTeam(props) {
                 filled='true'
                 type="text"
                 name="season"
-                value={props.season}
+                value={season}
                 required
-                onChange={(event) => props.setSeason(event.target.value)}
+                onChange={(event) => setSeason(event.target.value)}
               />
             </FormLabel>
           </Grid>
@@ -68,9 +96,9 @@ function CreateTeam(props) {
                 filled='true'
                 type="text"
                 name="yourPlayerNumber"
-                value={props.yourPlayerNumber}
+                value={yourPlayerNumber}
                 required
-                onChange={(event) => props.setYourPlayerNumber(event.target.value)}
+                onChange={(event) => setYourPlayerNumber(event.target.value)}
               />
             </FormLabel>
           </Grid>
